requirements.txt
astor==0.8.1
asttokens==2.4.1
atomicwrites==1.4.1
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
ax-platform==0.2.6
backcall==0.2.0
beautifulsoup4==4.12.3
# Editable install with no version control (BO-GP==0.0.1)
-e c:\users\xzt\desktop\code_sample
botorch==0.6.6
celer==0.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
ConfigSpace==0.7.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
download==0.3.5
emcee==3.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.49.0
formulaic==0.2.4
fsspec==2024.2.0
future==1.0.0
fvcore==0.1.5.post20210630
gdown==5.1.0
GPy==1.12.0
gpytorch==1.8.1
GraKeL==0.1.10
graphviz==0.20.1
hyperopt==0.2.7
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
iniconfig==2.0.0
interface-meta==1.3.0
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
-e git+https://github.com/ksehic/LassoBench.git@3e0d6a7e781aed50546443cf3ef3443a26ed4967#egg=LassoBench
libsvmdata==0.4.1
lifelines==0.27.8
lightgbm==3.2.1
linear-operator==0.4.0
llvmlite==0.39.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
more-itertools==10.2.0
mpmath==1.3.0
multipledispatch==1.0.0
nasbench301 @ git+https://github.com/automl/nasbench301@f5e8f308022e3260f4f5976a1bf355ec16fd93cd
nasbench_pytorch @ git+https://github.com/romulus0914/NASBench-PyTorch@37bc332bc6802e0375d8cca839d54f47524487fb
# Editable install with no version control (naslib==0.1.0)
-e c:\users\xzt\desktop\code_sample\naslib
nbformat==5.9.2
networkx==2.6.3
ngboost==0.3.11
numba==0.56.4
numpy==1.21.6
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
paramz==0.9.6
parso==0.8.3
pathvalidate==3.2.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.19.0
pluggy==0.13.1
portalocker==2.8.2
prompt-toolkit==3.0.43
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.7
pybnn==0.0.5
pyDOE==0.3.8
pydotz==1.5.1
Pygments==2.17.2
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.0
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.9.0.post0
pytorch-msssim==0.2.1
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.13.2
six==1.16.0
sobol-seq==0.2.0
soupsieve==2.5
sparse-ho @ https://github.com/QB3/sparse-ho/archive/master.zip#sha256=e9429a8237f808b5bf70f1f0d62d994e4d4b38c7c6e533ca1627a127d9c04368
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorwatch==0.9.1
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2023.7.10
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.61.1
traitlets==5.14.1
transforms3d==0.4.1
typeguard==2.13.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
widgetsnbextension==4.0.10
wrapt==1.16.0
xarray==2023.1.0
xgboost==1.4.2
yacs==0.1.8
zipp==3.17.0
